requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
beautifulsoup4==4.8.2
bs4==0.0.1
discord==1.0.1
discord-ext-menus==1.0.0a21+gd9965ec
discord.py==1.3.2
giphy-client==1.0.0
lxml==4.5.0
requests==2.22.0
requests-toolbelt==0.9.1
urllib3==1.25.8
websockets==8.1
Pillow==7.0.0
jikanpy==3.4.2
imgurpython==1.1.7
psutil==5.7.0
googletrans==2.4.0